:bug: fix(web): move window.__ENV injection out of the commercial-overlaid runtime module
The commercial build aliases `@/lib/commercial/runtime` to an analytics-only
provider (Magic-Resume-Commercial `next-config-overlay.mjs` webpack alias),
which stripped the `window.__ENV` injection — so the commercial image never got
the runtime apiOrigin and the browser fell back to http://localhost:3110 while
SSR/API routes (reading APP_API_ORIGIN) worked fine.

Move the injection into a dedicated, non-overlaid module
(`@/lib/commercial/runtime-env` → RuntimeEnvScript), render it from the root
layout, and slim CommercialRuntimeProvider back to a passthrough analytics slot.
The injection now survives both the OSS and commercial builds; the overlay only
swaps analytics.
